Armv6 & armv7 no need update SCTLR for alignment.
ARM processors cannot fetch/store information which is not naturally
aligned on the bus.
On 32-bit ARM processors, these non-aligned fetch/store instructions
will be emulated in software, which has a severe performance impact.

Now, when interrupt and exception disturbs the user-space, always
update SCTLR with cr_alignment.
In a ARMv7, the statements  bracketed by CONFIG_ALIGNMENT_TRAP
can need about 100 cycles, which impacts realtime interrupt scenario.

Considering in ARMv6 or ARMv7, cr_alignment and cr_no_alignment
are always same without A bit, they are initialized by clearing
A bit in alignment_init.

So, there is no need to update SCTLR if interrupt and exception
disturbs the user-space in vector_swi function or alignment_trap
macro, in the ARMv6 and ARMv7 chip.