futex-prevent-requeue-pi-on-same-futex
futex: Forbid  uaddr == uaddr2 in futex_requeue(..., requeue_pi=1)

If uaddr == uaddr2, then we have broken the rule of only requeueing
from a non-pi futex to a pi futex with this call.
If we attempt this, then dangling pointers may be left for rt_waiter
resulting in an exploitable condition.

Brings futex_requeue() into line with futex_wait_requeue_pi()
which performs the same check.