futex: Make lookup_pi_state more robust
The current implementation of lookup_pi_state has ambigous handling of
the TID value 0 in the user space futex.
We can get into the kernel even if the TID value is 0, because either
there is a stale waiters bit or the owner died bit is set or we are
called from the requeue_pi path or from user space just for fun.

The current code avoids an explicit sanity check for pid = 0 in case
that kernel internal state (waiters) are found for the user space
address.
This can lead to state leakage and worse under some circumstances.