Reduce large amounts of memory barrier
[get|put]_mems_allowed() is extremely xpensive and severely impacted
page allocator performance.

Fix no node to alloc memory when changing cpuset's mems"  wins a super
prize for the largest number of memory barriers entered into fast paths.

[get|put]_mems_allowed is incredibly heavy with pairs of full memory
barriers inserted into a number of hot paths.

This patch replaces the full memory barriers on both read and write
sides with a sequence counter with just read barriers on the fast path
side.  This is much cheaper on some architectures.

The main bulk of the patch is the retry logic if the nodemask changes
in a manner that can cause a false failure, but while updating the
nodemask, a check is made to see if a false failure is a risk.
If it is, the sequence number gets bumped and parallel allocators
will briefly stall while the nodemask update takes place.